+++ src/components/zen-drawer/zen-drawer.tsx
@@ -49,7 +49,12 @@ export class ZenDrawer {
   }
 
   componentDidLoad(): void {
-    this.opened ? showInstantly(this.drawer) : hideInstantly(this.drawer);
+    if (this.opened) {
+      showInstantly(this.drawer);
+    } else {
+      hideInstantly(this.drawer);
+      this.host.style.display = 'none';
+    }
   }
 
   render(): HTMLElement {
@@ -69,7 +74,7 @@ export class ZenDrawer {
           padding-bottom={this.paddingBottom}
           padding-left={this.paddingLeft}
         >
-          <ZenButton onClick={() => this.onCloseClicked()} class="close-btn" variant="tertiary">
+          <ZenButton onClick={() => this.onCloseClicked()} class="close-btn" variant="grey-text">
             <ZenIcon padding="sm" size="md" class="close-icon" icon={faArrowToRight}></ZenIcon>
           </ZenButton>
           <slot></slot>