+++ src/components/zen-tooltip/zen-tooltip.tsx
@@ -45,6 +45,9 @@ export class ZenTooltip {
   /** Set tooltip offset to target element */
   @Prop() readonly offset?: number = 10;
 
+  /** Set tooltip max width */
+  @Prop() readonly maxWidth: string = '22.5rem';
+
   /** Limit tooltips height and make content scroll  */
   @Prop() readonly maxHeight: string = 'none';
 
@@ -108,6 +111,7 @@ export class ZenTooltip {
           style={{
             color: this.color,
             'max-height': this.maxHeight,
+            'max-width': this.maxWidth,
           }}
           offset={{ x: 0, y: this.offset }}
           delay={this.delay}
@@ -124,7 +128,7 @@ export class ZenTooltip {
             </ZenText>
           )}
           <slot>
-            <ZenText size="sm" style={{ color: this.color }}>
+            <ZenText size="md" pale>
               {this.label}
             </ZenText>
           </slot>